+++ src/platform/core/steps/step-body/step-body.component.scss
