+++ src/platform/core/steps/step-header/step-header.component.html
